pyproject.toml
[project]
name = "codebase-intelligence-systems-trp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1.46",
    "jinja2>=3.1.6",
    "langchain-core>=1.2.18",
    "langgraph>=1.0.10",
    "litellm>=1.82.0",
    "networkx>=3.6.1",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "scipy>=1.17.1",
    "sqlglot>=29.0.1",
    "tree-sitter>=0.25.2",
    "tree-sitter-python>=0.25.0",
]

[dependency-groups]
dev = [
    "gitpython>=3.1.46",
    "jinja2>=3.1.6",
    "langchain-core>=1.2.18",
    "langgraph>=1.0.10",
    "litellm>=1.82.0",
    "networkx>=3.6.1",
    "openai>=2.26.0",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "sqlglot>=29.0.1",
    "tree-sitter>=0.25.2",
    "tree-sitter-python>=0.25.0",
]
